+++ tango/__init__.py
@@ -3,27 +3,43 @@ A Python library for choreographing your machine learning research.
 """
 
 __all__ = [
-    "Format",
+    "cleanup_cli",
     "DillFormat",
     "DillFormatIterator",
+    "execute_step_graph",
     "Executor",
+    "Format",
+    "initialize_cli",
     "JsonFormat",
     "JsonFormatIterator",
-    "SqliteDictFormat",
-    "Step",
-    "step",
-    "StepInfo",
-    "StepInfoSort",
-    "StepState",
-    "StepResources",
-    "StepCache",
-    "StepGraph",
+    "load_settings",
+    "prepare_executor",
+    "prepare_workspace",
     "Run",
     "RunInfo",
     "RunSort",
+    "SqliteDictFormat",
+    "Step",
+    "step",
+    "StepCache",
+    "StepGraph",
+    "StepInfo",
+    "StepInfoSort",
+    "StepResources",
+    "StepState",
+    "tango_cli",
     "Workspace",
 ]
 
+from .cli import (
+    cleanup_cli,
+    execute_step_graph,
+    initialize_cli,
+    load_settings,
+    prepare_executor,
+    prepare_workspace,
+    tango_cli,
+)
 from .executor import Executor
 from .format import (
     DillFormat,