+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Minor bugs in the `GSWorkspace()`.
 
+### Changed
+
+- Added CLI-style execution functions for experiments defined in Python.
+- Added `display()` to `ExecutorOutput` for producing a table that summarizes the run.
+
 ## [v1.3.0](https://github.com/allenai/tango/releases/tag/v1.3.0) - 2023-10-13
 
 ### Added